Make LocalProxy disavow the `__wrapped__` attribute
Allowing the `werkzeug.local.LocalProxy` to forward the request for the
`__wrapped__` attribute to the "RealSubject" breaks the `doctest` module
(which, as of Python 3.5, got more aggressive in looking for doctests
with `inspect.unwrap`).

This makes the `LocalProxy` handle the "__wrapped__" attribute on its
own (via call to a `super`), so this could be overridden on a case-by-
case basis, but makes the default behavior to **not** delegate that
attribute.